refactor(logic): eliminate diplomacy->turn import edge (Refs #3290)
Move the diplomacy-phase offer/decision value types and DiplomacyPhaseResult
out of turn/turn_resolution_result.dart into a new diplomacy-domain file
diplomacy/diplomacy_phase_result.dart. The turn orchestrator's sealed
TurnResolutionResult now imports + re-exports those types, so the dependency
runs one-way turn -> diplomacy and the four diplomacy resolvers no longer
import turn/. Adds (diplomacy, turn) to the logic domain import DAG forbidden
edges with enforcement + negative tests, and documents the eliminated edge in
SPEC/program/logic-package-split-phase0.md.

Behavior-preserving structural extraction; public barrel API unchanged via
re-export.
