+++ packages/colonizethis_logic/lib/src/diplomacy/diplomacy_resolver.dart
@@ -9,7 +9,7 @@ import 'package:colonizethis_logic/src/logging.dart';
 import 'package:colonizethis_models/colonizethis_models.dart';
 
 import '../constants.dart';
-import '../turn/turn_resolution_result.dart';
+import 'diplomacy_phase_result.dart';
 import 'package:colonizethis_world/src/world/faction_membership.dart';
 import 'package:colonizethis_world/src/world/province_lookup.dart';
