+++ packages/colonizethis_logic/lib/src/diplomacy/intervention_resolver.dart
@@ -8,7 +8,7 @@ import 'diplomacy_relation_lookup.dart';
 import '../combat/conflict_detection.dart';
 import '../constants.dart';
 import '../dossier/evidence_rules.dart';
-import '../turn/turn_resolution_result.dart';
+import 'diplomacy_phase_result.dart';
 import 'package:colonizethis_world/src/world/province_lookup.dart';
 import 'diplomacy_relation_updates.dart';
 import 'diplomacy_resolver.dart';
