+++ packages/colonizethis_logic/lib/src/diplomacy/overture_resolver.dart
@@ -3,7 +3,7 @@ import 'package:colonizethis_models/colonizethis_models.dart';
 
 import '../constants.dart';
 import '../dossier/evidence_rules.dart';
-import '../turn/turn_resolution_result.dart';
+import 'diplomacy_phase_result.dart';
 import 'diplomacy_resolver.dart';
 import 'overture_stage_helpers.dart';
