+++ packages/colonizethis_logic/lib/src/turn/turn_resolution_result.dart
@@ -6,244 +6,20 @@
 /// [TurnResolutionPendingIntervention] and blocks until [resumeTurnResolutionWithInterventionDecisions].
 /// When call to arms requires a human ally, resolution returns
 /// [TurnResolutionPendingCallToArms].
+///
+/// The diplomacy-domain offer/decision value types these variants carry
+/// ([OvertureOffer], [FtpOffer], [InterventionPrompt], [CallToArmsPending], and
+/// related decisions plus [DiplomacyPhaseResult]) live in the diplomacy domain
+/// (`diplomacy/diplomacy_phase_result.dart`) so the dependency runs one-way
+/// turn -> diplomacy (Refs #3290 Phase 0). They are re-exported here so existing
+/// consumers of this file keep seeing them unchanged.
 library;
 
 import 'package:colonizethis_models/colonizethis_models.dart';
 
-/// One overture offer awaiting the target's accept/reject decision.
-class OvertureOffer {
-  const OvertureOffer({
-    required this.offererGpId,
-    required this.targetFactionId,
-    required this.stage,
-  });
+import '../diplomacy/diplomacy_phase_result.dart';
 
-  final String offererGpId;
-  final String targetFactionId;
-  final OvertureStage stage;
-
-  @override
-  bool operator ==(Object other) =>
-      identical(this, other) ||
-      other is OvertureOffer &&
-          offererGpId == other.offererGpId &&
-          targetFactionId == other.targetFactionId &&
-          stage == other.stage;
-
-  @override
-  int get hashCode => Object.hash(offererGpId, targetFactionId, stage);
-}
-
-/// Target's decision for one overture offer.
-class OvertureDecision {
-  const OvertureDecision({
-    required this.offererGpId,
-    required this.targetFactionId,
-    required this.stage,
-    required this.accepted,
-  });
-
-  final String offererGpId;
-  final String targetFactionId;
-  final OvertureStage stage;
-  final bool accepted;
-
-  @override
-  bool operator ==(Object other) =>
-      identical(this, other) ||
-      other is OvertureDecision &&
-          offererGpId == other.offererGpId &&
-          targetFactionId == other.targetFactionId &&
-          stage == other.stage &&
-          accepted == other.accepted;
-
-  @override
-  int get hashCode =>
-      Object.hash(offererGpId, targetFactionId, stage, accepted);
-}
-
-/// Human intervention prompt after a GP declares war on a Minor or Tribe.
-/// SPEC/game/diplomacy.md § Intervention.
-class InterventionPrompt {
-  const InterventionPrompt({
-    required this.aggressorGpId,
-    required this.defenderMinorOrTribeId,
-    required this.interveningGpId,
-  });
-
-  final String aggressorGpId;
-  final String defenderMinorOrTribeId;
-  final String interveningGpId;
-
-  @override
-  bool operator ==(Object other) =>
-      identical(this, other) ||
-      other is InterventionPrompt &&
-          aggressorGpId == other.aggressorGpId &&
-          defenderMinorOrTribeId == other.defenderMinorOrTribeId &&
-          interveningGpId == other.interveningGpId;
-
-  @override
-  int get hashCode =>
-      Object.hash(aggressorGpId, defenderMinorOrTribeId, interveningGpId);
-}
-
-/// Player's intervention choice for one [InterventionPrompt].
-class InterventionDecision {
-  const InterventionDecision({
-    required this.aggressorGpId,
-    required this.defenderMinorOrTribeId,
-    required this.interveningGpId,
-    required this.choice,
-  });
-
-  final String aggressorGpId;
-  final String defenderMinorOrTribeId;
-  final String interveningGpId;
-  final InterventionChoice choice;
-
-  @override
-  bool operator ==(Object other) =>
-      identical(this, other) ||
-      other is InterventionDecision &&
-          aggressorGpId == other.aggressorGpId &&
-          defenderMinorOrTribeId == other.defenderMinorOrTribeId &&
-          interveningGpId == other.interveningGpId &&
-          choice == other.choice;
-
-  @override
-  int get hashCode => Object.hash(
-        aggressorGpId,
-        defenderMinorOrTribeId,
-        interveningGpId,
-        choice,
-      );
-}
-
-/// One call-to-arms prompt: ally [allyGpId] must choose to join defender [defenderGpId]
-/// against aggressor [aggressorGpId]. SPEC/game/diplomacy.md mutual defence.
-class CallToArmsPending {
-  const CallToArmsPending({
-    required this.allyGpId,
-    required this.defenderGpId,
-    required this.aggressorGpId,
-  });
-
-  final String allyGpId;
-  final String defenderGpId;
-  final String aggressorGpId;
-
-  @override
-  bool operator ==(Object other) =>
-      identical(this, other) ||
-      other is CallToArmsPending &&
-          allyGpId == other.allyGpId &&
-          defenderGpId == other.defenderGpId &&
-          aggressorGpId == other.aggressorGpId;
-
-  @override
-  int get hashCode => Object.hash(allyGpId, defenderGpId, aggressorGpId);
-}
-
-/// One FTP proposal awaiting the target GP's accept/reject decision.
-class FtpOffer {
-  const FtpOffer({
-    required this.proposerGpId,
-    required this.targetGpId,
-  });
-
-  final String proposerGpId;
-  final String targetGpId;
-
-  @override
-  bool operator ==(Object other) =>
-      identical(this, other) ||
-      other is FtpOffer &&
-          proposerGpId == other.proposerGpId &&
-          targetGpId == other.targetGpId;
-
-  @override
-  int get hashCode => Object.hash(proposerGpId, targetGpId);
-}
-
-/// Target GP's decision for one [FtpOffer].
-class FtpDecision {
-  const FtpDecision({
-    required this.proposerGpId,
-    required this.targetGpId,
-    required this.accepted,
-  });
-
-  final String proposerGpId;
-  final String targetGpId;
-  final bool accepted;
-
-  @override
-  bool operator ==(Object other) =>
-      identical(this, other) ||
-      other is FtpDecision &&
-          proposerGpId == other.proposerGpId &&
-          targetGpId == other.targetGpId &&
-          accepted == other.accepted;
-
-  @override
-  int get hashCode => Object.hash(proposerGpId, targetGpId, accepted);
-}
-
-/// Human ally's decision for one call to arms.
-class CallToArmsDecision {
-  const CallToArmsDecision({
-    required this.allyGpId,
-    required this.defenderGpId,
-    required this.aggressorGpId,
-    required this.accepted,
-  });
-
-  final String allyGpId;
-  final String defenderGpId;
-  final String aggressorGpId;
-  final bool accepted;
-
-  @override
-  bool operator ==(Object other) =>
-      identical(this, other) ||
-      other is CallToArmsDecision &&
-          allyGpId == other.allyGpId &&
-          defenderGpId == other.defenderGpId &&
-          aggressorGpId == other.aggressorGpId &&
-          accepted == other.accepted;
-
-  @override
-  int get hashCode =>
-      Object.hash(allyGpId, defenderGpId, aggressorGpId, accepted);
-}
-
-/// Result of the Diplomacy phase: complete or pending human input.
-class DiplomacyPhaseResult {
-  const DiplomacyPhaseResult(
-    this.game, {
-    this.pendingOvertures,
-    this.pendingFtpOffers,
-    this.pendingInterventions,
-    this.pendingCallToArms,
-  });
-
-  final Game game;
-  /// Non-null when phase suspended because an overture targets a human GP.
-  final List<OvertureOffer>? pendingOvertures;
-  /// Non-null when phase suspended because an FTP proposal targets a human GP.
-  final List<FtpOffer>? pendingFtpOffers;
-  /// Non-null when phase suspended for intervention choices (GP with embassy or purchased land).
-  final List<InterventionPrompt>? pendingInterventions;
-  /// Non-null when phase suspended because a human ally must accept/refuse call to arms.
-  final List<CallToArmsPending>? pendingCallToArms;
-
-  bool get isPending =>
-      (pendingOvertures != null && pendingOvertures!.isNotEmpty) ||
-      (pendingFtpOffers != null && pendingFtpOffers!.isNotEmpty) ||
-      (pendingInterventions != null && pendingInterventions!.isNotEmpty) ||
-      (pendingCallToArms != null && pendingCallToArms!.isNotEmpty);
-}
+export '../diplomacy/diplomacy_phase_result.dart';
 
 /// Sealed result of turn resolution: complete or pending human input.
 ///
