+++ test/check_logic_domain_import_dag_test.dart
@@ -139,6 +139,35 @@ void noop() {}
     expect(code, 1);
   });
 
+  test('diplomacy->turn edge is enforced and fully eliminated (Refs #3290)', () {
+    final forbidden = logicDomainImportForbiddenEdgesForTests();
+    expect(forbidden, contains('diplomacy->turn'));
+
+    final allowlist = logicDomainImportDagGrandfatherAllowlistForTests();
+    expect(allowlist.where((e) => e.startsWith('diplomacy->turn:')), isEmpty);
+  });
+
+  test('fails when a new forbidden diplomacy->turn import appears', () {
+    final temp = Directory.systemTemp.createTempSync('logic_dag_dipl_turn_');
+    addTearDown(() => temp.deleteSync(recursive: true));
+
+    File(
+      '${temp.path}/packages/colonizethis_logic/lib/src/diplomacy/bad_turn.dart',
+    )
+      ..createSync(recursive: true)
+      ..writeAsStringSync("""
+import '../turn/turn_resolution_result.dart';
+void noop() {}
+""");
+
+    final code = runCheckLogicDomainImportDag(
+      temp.path,
+      info: (_) {},
+      err: (_) {},
+    );
+    expect(code, 1);
+  });
+
   test(
     'world->combat and world->dossier leaf edges are enforced + eliminated',
     () {
